+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/mode/iconpack/conversion/IconPackConversionState.kt
@@ -75,4 +75,5 @@ enum class ValidationError {
     IconNameEmpty,
     IconNameContainsSpace,
     HasDuplicates,
+    HasCaseInsensitiveDuplicates,
 }